change state on label click
